Extend style customisation with custom function
Next to a chalk style, it can be very useful to extend the basic style
customisation with a function that returns text. For my use-case - a
unit test progress bar - I need to apply a style to the bar of the
progress bar that contains different colors at specific intervals. This
simple change allows for such customisations without breaking existing
behavior.