Make erasure of progressbar optional when ending
I would like to keep the progressbar - a unit test progress bar
containing passed, failed and skipped sections - on screen after
it reaches 100%, as an indicator to the user.
I do want to clean it up, i.e. call progress.end(), but with the option
of not erasing the latest drawn progressbar. I've currently implemented
it with a strict check for a boolean of type false, i.e. no falsy
support.